+++ Projecte/TileMap.h
@@ -37,7 +37,7 @@ public:
 	bool collisionMoveLeft(const glm::ivec2 &pos, const glm::ivec2 &size, const bool &bJumping, const bool ghost = false) const;
 	bool collisionMoveRight(const glm::ivec2 &pos, const glm::ivec2 &size, const bool &bJumping, const bool ghost = false) const;
 	bool collisionMoveUp(const glm::ivec2 &pos, const glm::ivec2 &size, const bool bJumping, const bool ghost = false) const;
-	bool collisionMoveDown(const glm::ivec2 &pos, const glm::ivec2 &size, int *posY, bool tileChanger, const bool ghost = false);
+	bool collisionMoveDown(const glm::ivec2 &pos, const glm::ivec2 &size, float *posY, bool tileChanger, const bool ghost = false);
 
 	bool collisionSpikes(const glm::ivec2 &pos, const glm::ivec2 &size);
